+++ Assets/Script/Tower/Tongue Tower/TongueTower.cs
@@ -14,7 +14,7 @@ namespace Tower.Tongue_Tower
             base.Start();
         }
         
-        void UpdateAttack(int newAttackDamage)
+        void UpdateAttack(float newAttackDamage)
         {
             TongueAttack.Damage = newAttackDamage;
             base.UpdateAttack(newAttackDamage);
